[0.3.1][Feature] Fully implemented DireccionController
